App: Transfer App funtions in app.js
